@@ -14,7 +14,7 @@
 	"devDependencies": {
 		"@sveltejs/adapter-vercel": "^5.4.7",
 		"@sveltejs/kit": "^2.8.1",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@sveltejs/vite-plugin-svelte": "^4.0.1",
 		"@types/eslint": "8.56.10",
 		"@types/qrcode": "^1.5.5",
 		"@typescript-eslint/eslint-plugin": "^7.18.0",
@@ -32,8 +32,8 @@
 		"prettier-plugin-svelte": "^3.2.8",
 		"prettier-plugin-tailwindcss": "^0.6.8",
 		"qrcode": "^1.5.4",
-		"svelte": "^4.2.19",
-		"svelte-check": "^3.8.6",
+		"svelte": "^5.2.2",
+		"svelte-check": "^4.0.9",
 		"sveltekit-superforms": "^2.20.1",
 		"tailwindcss": "^3.4.15",
 		"tslib": "^2.8.1",
@@ -42,7 +42,7 @@
 		"zod": "^3.23.8"
 	},
 	"type": "module",
-	"packageManager": "pnpm@9.1.0",
+	"packageManager": "pnpm@9.13.2",
 	"engines": {
 		"node": "^20 || ^22"
 	}
